Split commands into multiple scans.
Created multiple new commands to allow scans to be ran individually, but
maintaining backwards compatibility with the 'scan' command.
